+++ src/L2dotNET.GameService/network/serverpackets/ItemListItem.cs
@@ -5,13 +5,13 @@ namespace L2dotNET.GameService.Network.Serverpackets
         public int ObjectId;
         public int ItemId;
         public int Slot;
-        public long Count;
+        public int Count;
         public short Type2;
-        public int CType1;
+        public short CType1;
         public short Equip;
         public int Bodypart;
-        public int Enchant;
-        public int CType2;
+        public short Enchant;
+        public short CType2;
         public int Augment;
         public int Mana;
         public int TimeLeft;
